+++ processor/src/test/resources/expected/com/example/inject/ArezReactComponent_.java
@@ -62,6 +62,10 @@ abstract class ArezReactComponent_ extends ArezReactComponent {
     return componentConstructor;
   }
 
+  @Override
+  protected final void reportPropsChanged(@Nullable final BaseProps nextProps) {
+  }
+
   @JsType(
       isNative = true
   )