+++ processor/src/test/resources/input/com/example/arez/ComponentWithProp.java
@@ -1,0 +1,24 @@
+package com.example.arez;
+
+import javax.annotation.Nullable;
+import react4j.annotations.Prop;
+import react4j.annotations.ReactComponent;
+import react4j.arez.ReactArezComponent;
+import react4j.core.BaseContext;
+import react4j.core.BaseProps;
+import react4j.core.ReactNode;
+
+@ReactComponent
+abstract class ComponentWithProp
+  extends ReactArezComponent<BaseProps, BaseContext>
+{
+  @Prop
+  protected abstract String getValue();
+
+  @Nullable
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}